fix: #20 showing acc as int

+++ src/main/java/dev/osunolimits/models/FullUser.java
@@ -25,7 +25,7 @@ public class FullUser {
         private int pp;
         private int plays;
         private int playtime;
-        private int acc;
+        private double acc;
         private int max_combo;
         private long total_hits;
         private long replay_views;
